added error handling

+++ input.c
@@ -3,8 +3,8 @@
 #include <string.h>
 
 ssize_t get_line(char **line, FILE *fp) {
-    size_t buffer = 8;
-    size_t len = 0;
+    int buffer = 8;
+    int len = 0;
     *line = malloc(buffer);
     while (1) {
         if (fgets(*line + len, buffer - len, fp) == NULL) {
